SASCA refactor: connect pieces, start PyO3 interface
